+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 2.0.4+1
+* Run `Controller.configureRoutes` before mounting `@Expose` routes.
+* Make `Controller.configureServer` always return a `Future`.
+
 # 2.0.4
 * Prepare for Dart SDK change to `Stream<List<int>>` that are now
   `Stream<Uint8List>`.